add comment containing common anisotropy calcs
These are used in raw_accel to scale y faster than x. This compensates
for the smaller range in y and the fact that we move the most left and
right more easily than up and down.

There is a wee bit if trickery involved getting this to work in
raw_accel. If we just scale y up, it hits the limiter or runs off the
end of the table too quickly. Instead, we scale x down, then scale
everything back up so x is unchanged, but y is scaled up.