magnitude is not unique enough to warrant the complexity
The curves produced by this extra term are not different enough from
curves that can already be created.

In the end, this term mostly just serves to divide out below c more
strongly, but in a way that is slightly different because of the -1s.
The difference from just subtracting mag from nonlinearity doesn't
seem important enough to keep it in.