test negative nonlinearity
After correcting for putting the log in the mag term, the mag term
was negative, something I hadn't considered. I poked around a bit with
this and found if you multiply it by x, you get an s curve at the start.
We're starting to diverge from power law a bit, but this shape has a
bump right where tracking needs it, but it still flicks near and far.