new defaults after softplus adventure
I was messing with using softplus instead of x^m. It smooths out the
entry at the cost of extra sensitivity or, probably more correctly,
moving crossover, but that doesn't translate as well between curves.

After that, I realized I was too fast, so 18 it is, with new
corresponding limits.