testing pure exponential
I put my current curve and the pure exponential on desmos and made the
pure exponential fit as closely as I could. It was very close and I was
doing well, but the very beginning is too fast and I can't figure out
how to bring it down.

Back to exp by power again.