narrowing in on exp
This is very good. I spent a lot of time staring at the graph and what I learned was not what I figured. Trying to control this curve with just 3 inputs is challenging.

There is a correlation between nonlinearity and sensitivity. If you increase nonlinearity, there is a corrsponding increase to sensitivity that minimizes the difference in shape by minimizing the change in scale. I haven't figured it out yet, but at c=10, it was roughly between n and 2.5 n, up through about n=15. It may be the log, or log times linear.

I think something happens when the tangent line at c lines up with the linear graph. The curve just touches off the sensitivity line.

The only problems I have with this is it just doesn't sit down on my glass, and the tangent at 0 is large. I'm going to try crossfading from the line y=0.