new exp tune after figuring out a bit about limiter
The limiter affects transients. That momentary zip at the end of a flick
that comes from exponentiality is often more than 16 times normal
movement speed. When you take the limiter off, or set it very high, it
feels NOTICABLY freer and clearer, up until you're out past the sun.

I was setting the limiter so my 180s were always governed. This was as
low as 4 early on, but 8 most of the time. Now it is 32 and the
difference is stark. I have to explore a few of my old tunes with this
in mind.