add exponential_by_unit_logistic_log
This is the standard exponential multiplied by the logistic of the log.
This starts off the graph with a very smooth tangent near 0 with y
intercept near 0 as well.

Magnitude is used to control how quickly the logistic approaches 1.
At low m, the whole graph lays down with a wonderful tangent, but you
have to compensate for an extra dead spot if you lower it too much. At
high m, there is a pronounced, but smooth bump that happens well before
crossover. It counteracts the dead spot and you can tune it to pick up
just about when you are tracking up close.

It deviates from the power law by very little after the logistic hits
1.