tune exponential by logistic
This is not of the logistic log, just the logistic. With low crossover
to get into exp early, and low sense to flatten out that tangent, this
is one of better exponentials so far. Definitely the smoothest.

It just occurred to me that e^x is not a power law, and it's time to try
one out.