match log version from graph
I put the previous non-log tune up on the graph and made a log version
that matches as closely as possible, diverging only after crossover.

With crossover so low, though, you feel the difference, and it is
definitely slower, so I also lowered crossover by an arbitrary amount
to make it feel similar, but I don't have any way to quantify how much.