first serviceable log in a bit
I'm having a hell of a time tuning the logs because the power law
stuff doesn't tune the same (s = l, always), and I keep getting
interference from the limiter or the exp rush.

This is actually usable for me and the flicks are starting to land
again, but I'm only commiting this so I can keep it for a base.