last tune for power law log for a bit
I just cannot get this tuned. Initially it was because the limiter.
There is no way to make the limiter not break flicks. The linear
continuation of the last two points must continue and not be horizontal.
The diminishing rate the log gives is nice when you are just moving
smoothly, but it does not play well with flicking.

The floor is turning out to be a huge part of this. Those tunes that
feel good except the start/stop is sticky means the floor is too low.
Bumping it up to an (unscaled) constant really clears this up.