fantastic flipped ln
This takes the log, but scales it by half BEFORE running it through the
logisitic, which does two things. First, it really rounds out the shape,
making this smooth curve even smoother after the attack, but it also...
completely removes the attack? Without adjusting the tangent, this
curve starts out almost vertically. But if you adjust the taper using
magnitude, down all the way to .38, just enough taper comes back and
it has such a smooth log feel. This curve is great.