a nice, little cosine curve
I dialed this in on a graph to look a lot like my current gaussian of
ln tune. I don't like the range of tangents I could get in that part of
this curve, but that was mostly from trying to shoehorn the other curve
here. Still, it feels natural and very similar to my normal curve.