final tune of gaussian log
m=.6355 has the strange property that it doesn't feel like it has any
ramp, but it obviously does. m=.84 feels more accurate, but sluggish,
even though it has a higher final velocity. I looked at the graph to
really see what the differences were.

In .6355, the sense tangent is done before 5e-8. Even with oversampling
at density=16, this seems too much nuance to catch with a table-driven
method. That tangent is basically vertical. However, the tangent for
velocity is still very small, so there is no feeling of hitting a curb
yet. This means you don't really need an obvious, visible s on the sense
curve for the velocity to feel smooth.

.6355 feels so good for most usage. For flickscan and most of desktop,
it feels transparent. It is a little squirly for tracking, though, and
it shows some of the glass double click. So you need a little bit of an
s for practiacl usage.

I tried the midpoint between them, but it just seems worse than either.
How little tangent on sense is required?

Well, the graph shows a HUGE change between .6355 and .66. At 2/3 sense,
all of the tangent hook is below y = ox, so that tangent should still
be felt enough to stop at 0, but it still gets out of the way so fast
that you cant tell it's there without really looking.

2/3 looks like it might be it.