Trade some absolute scale for a hint of tangent.
This matches the tangent for the .6355 tune, but tilts down just a bit
of tangent at 0. It uses much less k, so the tangent at 1 is almost 1,
but this all had to come out of about a quarter of absolute scale.

It still feels plenty fast with the lower abs speed because it picks
up quickly enough that you don't overshoot velocity as often.